NOJIRA Add property placeholder support
Default values are in
ssp-data-importer-impl/src/main/resources/ssp-importer.default.properties

Use -Dssp.importer.configdir=/path/to/your/config dir to specify a
directory in which the app will look for overrides in a file named
ssp-importer.properties.

For demonstration purposes, this commit modifies the
RawItemCsvWriter to output the value of a dummy property. This can
be backed out once we're all agreed on the approach.

Also includes Spock tests to verify expected behavior.

Includes some logback configuration to suppress excessive Spring
logging when running tests. More work is still necessary to build out
logging config and overrides for the app proper.